transforming to io
