+++ examples/CMakeLists.txt
@@ -80,27 +80,28 @@ endif()
 
 find_package(Protobuf QUIET)
 if(PROTOBUF_FOUND)
-    set(proto_file "${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter/graph.pb.cc")
+    set(proto_file "${CMAKE_CURRENT_SOURCE_DIR}/../tiny_cnn/io/tensorflow/graph.pb.cc")
     if(EXISTS ${proto_file})
         file(GLOB proto_src "${CMAKE_CURRENT_SOURCE_DIR}/../tensorflow/core/framework/*.pb.cc")
         add_executable(example_tensorflow_converter
-                       ${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter/example_tensorflow_parser.cpp ${proto_src})
+                       ${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter/tensorflow_parser.cpp ${proto_src})
         #              ${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter/example_tensorflow_parser.cpp ${proto_file} ${proto_src})
         target_link_libraries(example_tensorflow_converter
             ${project_library_target_name}
             ${PROTOBUF_LIBRARIES} ${REQUIRED_LIBRARIES})
     elseif(EXISTS ${PROTOBUF_PROTOC_EXECUTABLE})
         message(STATUS "Found PROTOBUF Compiler: ${PROTOBUF_PROTOC_EXECUTABLE}")
-        if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter/graph.proto)
+        if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/../tiny_cnn/io/tensorflow/graph.proto)
             PROTOBUF_GENERATE_CPP(PROTO_SRCS PROTO_HDRS
-                ${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter/graph.proto)
+                ${CMAKE_CURRENT_SOURCE_DIR}/../tiny_cnn/io/tensorflow/graph.proto)
+            file(GLOB proto_src "${CMAKE_CURRENT_SOURCE_DIR}/../tensorflow/core/framework/*.pb.cc")
             if(EXISTS ${PROTO_HDRS})
-                file(COPY ${PROTO_SRCS} DESTINATION "${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter")
-                file(COPY ${PROTO_HDRS} DESTINATION "${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter")
+                file(COPY ${PROTO_SRCS} DESTINATION "${CMAKE_CURRENT_SOURCE_DIR}/../tiny_cnn/io/tensorflow")
+                file(COPY ${PROTO_HDRS} DESTINATION "${CMAKE_CURRENT_SOURCE_DIR}/../tiny_cnn/io/tensorflow")
             endif()
             add_executable(example_tensorflow_converter
-                           ${CMAKE_CURRENT_SOURCE_DIR}/examples/tensorflow_converter/example_tensorflow_parser.cpp
-                           ${PROTO_SRCS} ${PROTO_HDRS})
+                           ${CMAKE_CURRENT_SOURCE_DIR}/../examples/tensorflow_converter/tensorflow_parser.cpp
+                           ${proto_src})
             target_link_libraries(example_tensorflow_converter
                 ${project_library_target_name}
                 ${PROTOBUF_LIBRARIES} ${REQUIRED_LIBRARIES})
